fix: Address remaining review feedback on PR #1667
- Issue #3: Document hard tool enforcement mechanism in architecture.md
- Issue #4: Codex now rejects disallowed_tools as unsupported instead of
  silently dropping it
- Issue #5: Move disallowed_tools tests from inline mod to launch_tests.rs
- Issue #6: Add comment explaining nudge failure safety (hard restrictions
  still apply via --disallowedTools)
- Issue #7: Update build_claude_headless_args doc comment to mention
  --disallowedTools flag
