+++ docs/architecture.md
@@ -243,6 +243,8 @@ Note: `route_mentions()` is intentionally disabled for topic channels — user `
 
 **Domain context from notes:** The `{domain_context}` variable is populated by `load_channel_notes()` in `src/channel.rs`, which reads all `.md` files from `channels/<name>/notes/`, concatenates them with filename-derived headers, and caps total size at 100 KB. This is called at all 4 channel lead spawn sites (3 in `effects.rs`, 1 in `cli/lead.rs`) so that channel leads always start with their accumulated domain knowledge. Insight nudges include a reminder to save important knowledge to notes, completing the feedback loop.
 
+**Hard tool restrictions:** Channel leads have code-modification tools (`Edit`, `Write`, `NotebookEdit`) blocked at the CLI level via `--disallowedTools`, enforced by `channel_lead_disallowed_tools()` in `src/launch.rs`. This is a hard enforcement mechanism that the LLM cannot bypass — the Claude Code CLI rejects these tool calls before execution. `Bash` is intentionally *not* blocked because channel leads need it for coordination commands (`midtown task create`, `midtown channel post`, etc.). The soft system prompt instruction in `channel-lead.md` ("Do NOT use Edit, Write, or Bash to modify code") remains as behavioral guidance for Bash usage. When channel leads fork into thread-specific sessions, the fork inherits the same `disallowed_tools` via `create_fork_session()` in `rpc_session.rs` (conditional on the parent being a channel lead). Non-channel-lead forks (e.g., coworker forks) receive no tool restrictions. The `HeadlessConfig.disallowed_tools` field carries the restriction list through `build_claude_headless_args()` in `platform.rs`. Note: Codex does not support `disallowed_tools` and will return an error if a channel lead is configured to use the Codex provider.
+
 **Coworker guidance:** Coworkers are instructed to `@{channel-name}` (e.g., `@daemon-core`) for domain questions and to reserve `@{project_name}` for coordination, task, and priority questions.
 
 ### Forked Sessions (Thread-Specific Channel Leads)
