+++ src/launch.rs
@@ -1309,75 +1309,6 @@ mod tests {
             "System prompt should reference the channel name"
         );
     }
-
-    #[test]
-    fn test_channel_lead_disallowed_tools_contains_code_modification_tools() {
-        let tools = channel_lead_disallowed_tools();
-        assert!(tools.contains(&"Edit".to_string()));
-        assert!(tools.contains(&"Write".to_string()));
-        assert!(tools.contains(&"NotebookEdit".to_string()));
-        // Bash is intentionally NOT blocked — channel leads need it for
-        // coordination commands (midtown task create, midtown channel post, etc.)
-        assert!(!tools.contains(&"Bash".to_string()));
-    }
-
-    #[test]
-    fn test_channel_lead_headless_config_has_disallowed_tools() {
-        let config = LaunchConfig::channel_lead("auth", "myrepo", SessionMode::Fresh, "");
-        let headless = config.to_headless_config("midtown");
-        assert!(
-            !headless.disallowed_tools.is_empty(),
-            "Channel lead should have disallowed tools"
-        );
-        assert!(
-            headless.disallowed_tools.contains(&"Edit".to_string()),
-            "Channel lead should disallow Edit"
-        );
-        assert!(
-            headless.disallowed_tools.contains(&"Write".to_string()),
-            "Channel lead should disallow Write"
-        );
-        assert!(
-            !headless.disallowed_tools.contains(&"Bash".to_string()),
-            "Channel lead should NOT disallow Bash (needed for midtown CLI commands)"
-        );
-        assert!(
-            headless
-                .disallowed_tools
-                .contains(&"NotebookEdit".to_string()),
-            "Channel lead should disallow NotebookEdit"
-        );
-    }
-
-    #[test]
-    fn test_coworker_headless_config_has_no_disallowed_tools() {
-        let config = LaunchConfig::coworker("park", "myrepo", SessionMode::Fresh, None);
-        let headless = config.to_headless_config("midtown");
-        assert!(
-            headless.disallowed_tools.is_empty(),
-            "Coworker should not have disallowed tools"
-        );
-    }
-
-    #[test]
-    fn test_reviewer_headless_config_has_no_disallowed_tools() {
-        let config = LaunchConfig::reviewer("york", "myrepo", 42, 0, AuthProvider::Claude);
-        let headless = config.to_headless_config("midtown");
-        assert!(
-            headless.disallowed_tools.is_empty(),
-            "Reviewer should not have disallowed tools"
-        );
-    }
-
-    #[test]
-    fn test_lead_headless_config_has_no_disallowed_tools() {
-        let config = LaunchConfig::lead("myrepo", None);
-        let headless = config.to_headless_config("midtown");
-        assert!(
-            headless.disallowed_tools.is_empty(),
-            "Lead should not have disallowed tools"
-        );
-    }
 }
 
 #[path = "launch_tests.rs"]
